requirements.txt
google-adk
distro
langchain
langchain-community
litellm
google-cloud-aiplatform
google-cloud-storage
google-cloud-bigquery
google-cloud-secret-manager
google-api-python-client
google-generativeai
youtube-transcript-api
markdown
scipy
# scikit-learn  # Temporarily disabled for deployment
pandas
numpy
# lightgbm  # Temporarily disabled for deployment
langchain-google-vertexai
# neo4j  # Disabled - using JSON files instead
fastapi
uvicorn
python-dotenv
pydantic==2.11.7
# unstructured  # Temporarily disabled for deployment
joblib
yfinance
praw
textblob
vaderSentiment
tweepy
stripe
sqlalchemy
databases[sqlite]
asyncpg
databases[postgresql]
passlib[bcrypt]
python-multipart
jinja2
itsdangerous
celery
flask
# redis  # Disabled - not needed for basic deployment
matplotlib
